fix: setting: react-query-devtools 세팅, selectorbox 버그 수정, User헤더 Info 컴포넌트 분리
