+++ package.json
@@ -10,6 +10,7 @@
     "@emotion/react": "^11.10.5",
     "@emotion/styled": "^11.10.5",
     "@tanstack/react-query": "^4.16.1",
+    "@tanstack/react-query-devtools": "^4.20.4",
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
@@ -79,8 +80,8 @@
     "eslint-plugin-jsx-a11y": "^6.6.1",
     "eslint-plugin-react": "^7.31.11",
     "eslint-plugin-react-hooks": "^4.6.0",
-    "husky": "^8.0.0",
     "framer-motion": "^8.0.2",
+    "husky": "^8.0.0",
     "lint-staged": "^13.0.3",
     "prettier": "^2.8.0"
   }
